# api/openapi/dashboard.v1.yaml
openapi: 3.0.3
info:
  title: screpdb Dashboard API
  version: 1.0.0
paths:
...
  /api/custom/ingest/settings:
    get:
      operationId: getIngestSettings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
    put:
      operationId: updateIngestSettings
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIngestSettingsRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
...
  /api/custom/aliases:
    get:
      operationId: listAliases
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
    put:
      operationId: importAliases
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportAliasesRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/custom/aliases/entry:
    put:
      operationId: upsertAliasEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertAliasEntryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/custom/aliases/{id}:
    delete:
      operationId: deleteAliasEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/custom/global-replay-filter:
    get:
      operationId: getGlobalReplayFilterConfig
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
    put:
      operationId: updateGlobalReplayFilterConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGlobalReplayFilterConfigRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
...
  /api/games:
    get:
      operationId: gamesList
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: player
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: map
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: duration
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: featuring
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: matchup
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: map_kind
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/games/{replayID}:
    parameters:
      - $ref: "#/components/parameters/replayID"
    get:
      operationId: gameDetail
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/games/{replayID}/ask:
    parameters:
      - $ref: "#/components/parameters/replayID"
    post:
      operationId: gameAsk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/games/{replayID}/see:
    parameters:
      - $ref: "#/components/parameters/replayID"
    post:
      operationId: gameSee
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/players:
    get:
      operationId: playersList
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: only_5_plus
          in: query
          required: false
          schema:
            type: string
        - name: sort_by
          in: query
          required: false
          schema:
            type: string
            enum: [name, race, games, apm, last_played]
        - name: sort_dir
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
        - name: last_played
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
...
  /api/players/insights/unit-production-cadence:
    get:
      operationId: playersUnitCadence
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
        - name: min_games
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
...
  /api/players/{playerKey}:
    parameters:
      - $ref: "#/components/parameters/playerKey"
    get:
      operationId: playerDetail
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/players/{playerKey}/recent-games:
    parameters:
      - $ref: "#/components/parameters/playerKey"
    get:
      operationId: playerRecentGames
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/players/{playerKey}/chat-summary:
    parameters:
      - $ref: "#/components/parameters/playerKey"
    get:
      operationId: playerChatSummary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/players/{playerKey}/insight:
    parameters:
      - $ref: "#/components/parameters/playerKey"
      - name: type
        in: query
        required: false
        schema:
          type: string
    get:
      operationId: playerInsight
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/players/{playerKey}/outliers:
    parameters:
      - $ref: "#/components/parameters/playerKey"
    get:
      operationId: playerOutliers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/players/{playerKey}/insights/apm-histogram:
    parameters:
      - $ref: "#/components/parameters/playerKey"
    get:
      operationId: playerApmHistogram
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/players/{playerKey}/insights/first-unit-delay:
    parameters:
      - $ref: "#/components/parameters/playerKey"
    get:
      operationId: playerDelayInsight
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/players/{playerKey}/insights/unit-production-cadence:
    parameters:
      - $ref: "#/components/parameters/playerKey"
      - name: filter
        in: query
        required: false
        schema:
          type: string
    get:
      operationId: playerUnitCadence
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
  /api/players/{playerKey}/ask:
    parameters:
      - $ref: "#/components/parameters/playerKey"
    post:
      operationId: playerAsk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericValue"
...
components:
  parameters:
    replayID:
      name: replayID
      in: path
      required: true
      schema:
        type: integer
        format: int64
    playerKey:
      name: playerKey
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    UpdateIngestSettingsRequest:
      type: object
      additionalProperties: false
      properties:
        input_dir:
          type: string
    AliasImportEntry:
      type: object
      additionalProperties: false
      required: [battle_tag]
      properties:
        aurora_id:
          type: integer
          format: int64
        battle_tag:
          type: string
    ImportAliasesRequest:
      type: object
      additionalProperties: false
      required: [aliases]
      properties:
        aliases:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: "#/components/schemas/AliasImportEntry"
    UpsertAliasEntryRequest:
      type: object
      additionalProperties: false
      required: [canonical_alias, battle_tag]
      properties:
        canonical_alias:
          type: string
        battle_tag:
          type: string
        aurora_id:
          type: integer
          format: int64
        source:
          type: string
          enum: [manual, imported, you]
    UpdateGlobalReplayFilterConfigRequest:
      type: object
      additionalProperties: false
      required:
        - game_types
        - exclude_short_games
        - exclude_computers
        - map_kinds
      properties:
        game_types:
          type: array
          items:
            type: string
            enum: [top_vs_bottom, melee, one_on_one, free_for_all]
        exclude_short_games:
          type: boolean
        exclude_computers:
          type: boolean
        map_kinds:
          type: array
          description: |
            Map types to include. UMS replays are always excluded globally
            (auto-discarded at ingest), so 'ums' is not a valid value here.
          items:
            type: string
            enum: [regular, money]
        compiled_replays_filter_sql:
          type: string
          nullable: true
    AskRequest:
      type: object
      additionalProperties: false
      required: [question]
      properties:
        question:
          type: string
    GenericObject:
      type: object
      additionalProperties: true
    GenericValue:
      oneOf:
        - $ref: "#/components/schemas/GenericObject"
        - type: array
          items:
            oneOf:
              - $ref: "#/components/schemas/GenericObject"
              - type: string
              - type: number
              - type: integer
              - type: boolean
        - type: string
        - type: number
        - type: integer
        - type: boolean