# internal/dashboard/openapi_handlers_gen.go
package dashboard

import (
	"context"

	"encoding/json"
	"net/http"

	"github.com/marianogappa/screpdb/internal/dashboard/apigen"
)

// Code generated by gen_openapi_bridge. DO NOT EDIT.
type ListAliasesJSONResponse struct {
	Payload any
}

func (response ListAliasesJSONResponse) VisitListAliasesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) ListAliases(ctx context.Context, request apigen.ListAliasesRequestObject) (apigen.ListAliasesResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.ListAliases, func(value any) apigen.ListAliasesResponseObject { return ListAliasesJSONResponse{Payload: value} })
}

type ImportAliasesJSONResponse struct {
	Payload any
}

func (response ImportAliasesJSONResponse) VisitImportAliasesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) ImportAliases(ctx context.Context, request apigen.ImportAliasesRequestObject) (apigen.ImportAliasesResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.ImportAliases, func(value any) apigen.ImportAliasesResponseObject { return ImportAliasesJSONResponse{Payload: value} })
}

type UpsertAliasEntryJSONResponse struct {
	Payload any
}

func (response UpsertAliasEntryJSONResponse) VisitUpsertAliasEntryResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) UpsertAliasEntry(ctx context.Context, request apigen.UpsertAliasEntryRequestObject) (apigen.UpsertAliasEntryResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.UpsertAliasEntry, func(value any) apigen.UpsertAliasEntryResponseObject {
		return UpsertAliasEntryJSONResponse{Payload: value}
	})
}

type DeleteAliasEntryJSONResponse struct {
	Payload any
}

func (response DeleteAliasEntryJSONResponse) VisitDeleteAliasEntryResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) DeleteAliasEntry(ctx context.Context, request apigen.DeleteAliasEntryRequestObject) (apigen.DeleteAliasEntryResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.DeleteAliasEntry, func(value any) apigen.DeleteAliasEntryResponseObject {
		return DeleteAliasEntryJSONResponse{Payload: value}
	})
}

type GetGlobalReplayFilterConfigJSONResponse struct {
	Payload any
}

func (response GetGlobalReplayFilterConfigJSONResponse) VisitGetGlobalReplayFilterConfigResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) GetGlobalReplayFilterConfig(ctx context.Context, request apigen.GetGlobalReplayFilterConfigRequestObject) (apigen.GetGlobalReplayFilterConfigResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.GetGlobalReplayFilterConfig, func(value any) apigen.GetGlobalReplayFilterConfigResponseObject {
		return GetGlobalReplayFilterConfigJSONResponse{Payload: value}
	})
}

type UpdateGlobalReplayFilterConfigJSONResponse struct {
	Payload any
}

func (response UpdateGlobalReplayFilterConfigJSONResponse) VisitUpdateGlobalReplayFilterConfigResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) UpdateGlobalReplayFilterConfig(ctx context.Context, request apigen.UpdateGlobalReplayFilterConfigRequestObject) (apigen.UpdateGlobalReplayFilterConfigResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.UpdateGlobalReplayFilterConfig, func(value any) apigen.UpdateGlobalReplayFilterConfigResponseObject {
		return UpdateGlobalReplayFilterConfigJSONResponse{Payload: value}
	})
}

type GetGlobalReplayFilterOptionsJSONResponse struct {
	Payload any
}

func (response GetGlobalReplayFilterOptionsJSONResponse) VisitGetGlobalReplayFilterOptionsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) GetGlobalReplayFilterOptions(ctx context.Context, request apigen.GetGlobalReplayFilterOptionsRequestObject) (apigen.GetGlobalReplayFilterOptionsResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.GetGlobalReplayFilterOptions, func(value any) apigen.GetGlobalReplayFilterOptionsResponseObject {
		return GetGlobalReplayFilterOptionsJSONResponse{Payload: value}
	})
}

type IngestJSONResponse struct {
	Payload any
}

func (response IngestJSONResponse) VisitIngestResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) Ingest(ctx context.Context, request apigen.IngestRequestObject) (apigen.IngestResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.Ingest, func(value any) apigen.IngestResponseObject { return IngestJSONResponse{Payload: value} })
}

func (a *openAPIStrictAdapter) IngestLogs(ctx context.Context, request apigen.IngestLogsRequestObject) (apigen.IngestLogsResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.IngestLogs, func(_ any) apigen.IngestLogsResponseObject { return apigen.IngestLogs101Response{} })
}

type GetIngestSettingsJSONResponse struct {
	Payload any
}

func (response GetIngestSettingsJSONResponse) VisitGetIngestSettingsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) GetIngestSettings(ctx context.Context, request apigen.GetIngestSettingsRequestObject) (apigen.GetIngestSettingsResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.GetIngestSettings, func(value any) apigen.GetIngestSettingsResponseObject {
		return GetIngestSettingsJSONResponse{Payload: value}
	})
}

type UpdateIngestSettingsJSONResponse struct {
	Payload any
}

func (response UpdateIngestSettingsJSONResponse) VisitUpdateIngestSettingsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) UpdateIngestSettings(ctx context.Context, request apigen.UpdateIngestSettingsRequestObject) (apigen.UpdateIngestSettingsResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.UpdateIngestSettings, func(value any) apigen.UpdateIngestSettingsResponseObject {
		return UpdateIngestSettingsJSONResponse{Payload: value}
	})
}

type GetStaleReplaysCountJSONResponse struct {
	Payload any
}

func (response GetStaleReplaysCountJSONResponse) VisitGetStaleReplaysCountResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) GetStaleReplaysCount(ctx context.Context, request apigen.GetStaleReplaysCountRequestObject) (apigen.GetStaleReplaysCountResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.GetStaleReplaysCount, func(value any) apigen.GetStaleReplaysCountResponseObject {
		return GetStaleReplaysCountJSONResponse{Payload: value}
	})
}

type GamesListJSONResponse struct {
	Payload any
}

func (response GamesListJSONResponse) VisitGamesListResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) GamesList(ctx context.Context, request apigen.GamesListRequestObject) (apigen.GamesListResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.GamesList, func(value any) apigen.GamesListResponseObject { return GamesListJSONResponse{Payload: value} })
}

type GameDetailJSONResponse struct {
	Payload any
}

func (response GameDetailJSONResponse) VisitGameDetailResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) GameDetail(ctx context.Context, request apigen.GameDetailRequestObject) (apigen.GameDetailResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.GameDetail, func(value any) apigen.GameDetailResponseObject { return GameDetailJSONResponse{Payload: value} })
}

type GameAskJSONResponse struct {
	Payload any
}

func (response GameAskJSONResponse) VisitGameAskResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) GameAsk(ctx context.Context, request apigen.GameAskRequestObject) (apigen.GameAskResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.GameAsk, func(value any) apigen.GameAskResponseObject { return GameAskJSONResponse{Payload: value} })
}

type GameSeeJSONResponse struct {
	Payload any
}

func (response GameSeeJSONResponse) VisitGameSeeResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) GameSee(ctx context.Context, request apigen.GameSeeRequestObject) (apigen.GameSeeResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.GameSee, func(value any) apigen.GameSeeResponseObject { return GameSeeJSONResponse{Payload: value} })
}

type HealthcheckJSONResponse struct {
	Payload any
}

func (response HealthcheckJSONResponse) VisitHealthcheckResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) Healthcheck(ctx context.Context, request apigen.HealthcheckRequestObject) (apigen.HealthcheckResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.Healthcheck, func(value any) apigen.HealthcheckResponseObject { return HealthcheckJSONResponse{Payload: value} })
}

type PlayerColorsJSONResponse struct {
	Payload any
}

func (response PlayerColorsJSONResponse) VisitPlayerColorsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerColors(ctx context.Context, request apigen.PlayerColorsRequestObject) (apigen.PlayerColorsResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerColors, func(value any) apigen.PlayerColorsResponseObject { return PlayerColorsJSONResponse{Payload: value} })
}

type PlayersListJSONResponse struct {
	Payload any
}

func (response PlayersListJSONResponse) VisitPlayersListResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayersList(ctx context.Context, request apigen.PlayersListRequestObject) (apigen.PlayersListResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayersList, func(value any) apigen.PlayersListResponseObject { return PlayersListJSONResponse{Payload: value} })
}

type PlayersApmHistogramJSONResponse struct {
	Payload any
}

func (response PlayersApmHistogramJSONResponse) VisitPlayersApmHistogramResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayersApmHistogram(ctx context.Context, request apigen.PlayersApmHistogramRequestObject) (apigen.PlayersApmHistogramResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayersApmHistogram, func(value any) apigen.PlayersApmHistogramResponseObject {
		return PlayersApmHistogramJSONResponse{Payload: value}
	})
}

type PlayersDelayHistogramJSONResponse struct {
	Payload any
}

func (response PlayersDelayHistogramJSONResponse) VisitPlayersDelayHistogramResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayersDelayHistogram(ctx context.Context, request apigen.PlayersDelayHistogramRequestObject) (apigen.PlayersDelayHistogramResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayersDelayHistogram, func(value any) apigen.PlayersDelayHistogramResponseObject {
		return PlayersDelayHistogramJSONResponse{Payload: value}
	})
}

type PlayersUnitCadenceJSONResponse struct {
	Payload any
}

func (response PlayersUnitCadenceJSONResponse) VisitPlayersUnitCadenceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayersUnitCadence(ctx context.Context, request apigen.PlayersUnitCadenceRequestObject) (apigen.PlayersUnitCadenceResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayersUnitCadence, func(value any) apigen.PlayersUnitCadenceResponseObject {
		return PlayersUnitCadenceJSONResponse{Payload: value}
	})
}

type PlayersViewportMultitaskingJSONResponse struct {
	Payload any
}

func (response PlayersViewportMultitaskingJSONResponse) VisitPlayersViewportMultitaskingResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayersViewportMultitasking(ctx context.Context, request apigen.PlayersViewportMultitaskingRequestObject) (apigen.PlayersViewportMultitaskingResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayersViewportMultitasking, func(value any) apigen.PlayersViewportMultitaskingResponseObject {
		return PlayersViewportMultitaskingJSONResponse{Payload: value}
	})
}

type PlayerDetailJSONResponse struct {
	Payload any
}

func (response PlayerDetailJSONResponse) VisitPlayerDetailResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerDetail(ctx context.Context, request apigen.PlayerDetailRequestObject) (apigen.PlayerDetailResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerDetail, func(value any) apigen.PlayerDetailResponseObject { return PlayerDetailJSONResponse{Payload: value} })
}

type PlayerAskJSONResponse struct {
	Payload any
}

func (response PlayerAskJSONResponse) VisitPlayerAskResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerAsk(ctx context.Context, request apigen.PlayerAskRequestObject) (apigen.PlayerAskResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerAsk, func(value any) apigen.PlayerAskResponseObject { return PlayerAskJSONResponse{Payload: value} })
}

type PlayerChatSummaryJSONResponse struct {
	Payload any
}

func (response PlayerChatSummaryJSONResponse) VisitPlayerChatSummaryResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerChatSummary(ctx context.Context, request apigen.PlayerChatSummaryRequestObject) (apigen.PlayerChatSummaryResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerChatSummary, func(value any) apigen.PlayerChatSummaryResponseObject {
		return PlayerChatSummaryJSONResponse{Payload: value}
	})
}

type PlayerInsightJSONResponse struct {
	Payload any
}

func (response PlayerInsightJSONResponse) VisitPlayerInsightResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerInsight(ctx context.Context, request apigen.PlayerInsightRequestObject) (apigen.PlayerInsightResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerInsight, func(value any) apigen.PlayerInsightResponseObject { return PlayerInsightJSONResponse{Payload: value} })
}

type PlayerApmHistogramJSONResponse struct {
	Payload any
}

func (response PlayerApmHistogramJSONResponse) VisitPlayerApmHistogramResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerApmHistogram(ctx context.Context, request apigen.PlayerApmHistogramRequestObject) (apigen.PlayerApmHistogramResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerApmHistogram, func(value any) apigen.PlayerApmHistogramResponseObject {
		return PlayerApmHistogramJSONResponse{Payload: value}
	})
}

type PlayerDelayInsightJSONResponse struct {
	Payload any
}

func (response PlayerDelayInsightJSONResponse) VisitPlayerDelayInsightResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerDelayInsight(ctx context.Context, request apigen.PlayerDelayInsightRequestObject) (apigen.PlayerDelayInsightResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerDelayInsight, func(value any) apigen.PlayerDelayInsightResponseObject {
		return PlayerDelayInsightJSONResponse{Payload: value}
	})
}

type PlayerUnitCadenceJSONResponse struct {
	Payload any
}

func (response PlayerUnitCadenceJSONResponse) VisitPlayerUnitCadenceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerUnitCadence(ctx context.Context, request apigen.PlayerUnitCadenceRequestObject) (apigen.PlayerUnitCadenceResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerUnitCadence, func(value any) apigen.PlayerUnitCadenceResponseObject {
		return PlayerUnitCadenceJSONResponse{Payload: value}
	})
}

type PlayerOutliersJSONResponse struct {
	Payload any
}

func (response PlayerOutliersJSONResponse) VisitPlayerOutliersResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerOutliers(ctx context.Context, request apigen.PlayerOutliersRequestObject) (apigen.PlayerOutliersResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerOutliers, func(value any) apigen.PlayerOutliersResponseObject { return PlayerOutliersJSONResponse{Payload: value} })
}

type PlayerRecentGamesJSONResponse struct {
	Payload any
}

func (response PlayerRecentGamesJSONResponse) VisitPlayerRecentGamesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) PlayerRecentGames(ctx context.Context, request apigen.PlayerRecentGamesRequestObject) (apigen.PlayerRecentGamesResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.PlayerRecentGames, func(value any) apigen.PlayerRecentGamesResponseObject {
		return PlayerRecentGamesJSONResponse{Payload: value}
	})
}

type ScrepColorsJSONResponse struct {
	Payload any
}

func (response ScrepColorsJSONResponse) VisitScrepColorsResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	return json.NewEncoder(w).Encode(response.Payload)
}

func (a *openAPIStrictAdapter) ScrepColors(ctx context.Context, request apigen.ScrepColorsRequestObject) (apigen.ScrepColorsResponseObject, error) {
	return responseFromPayload(ctx, request, a.service.ScrepColors, func(value any) apigen.ScrepColorsResponseObject { return ScrepColorsJSONResponse{Payload: value} })
}